сделано два афинных преобразования

+++ src/main/java/laba_1/MyPoint2D.java
@@ -1,20 +1,20 @@
 package laba_1;
 
 public class MyPoint2D {
-    private int x, y;
+    private double x, y;
 
     public MyPoint2D() { }
 
-    public MyPoint2D(int x, int y) {
+    public MyPoint2D(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 }
